Add files via upload
Data pre-processing using R and Python scripts, we initially load all the CSV files into the workbook and combine the data into a single dataframe and clean it. Later we renamed the columns with shorter and more intuitive names and separated the date and time column into two separate columns. Using Python, we performed a statistical pre-analysis with Pearson, Kendal and Speraman correlations.